pyproject.toml
[project]
name = "spy-trader"
version = "0.1.0"
description = "Autonomous Elder Triple Screen paper-trading agent for SPY/SH on Alpaca."
requires-python = ">=3.11,<3.12"
dependencies = [
  "alpaca-py>=0.30",
  "pandas>=2.2",
  "pyarrow>=16.0",
  "pandas-market-calendars>=4.4",
  "pytz>=2024.1",
  "httpx>=0.27",
  "anthropic>=0.40",
  "jinja2>=3.1",
  "python-dotenv>=1.0",
  "pydantic>=2.7",
]

[project.scripts]
spy-trader = "spy_trader.cli:main"

[dependency-groups]
dev = [
  "pytest>=8.2",
  "pytest-mock>=3.14",
  "pytest-cov>=5.0",
  "ruff>=0.5",
  "mypy>=1.10",
  "respx>=0.21",
]

[tool.uv]
package = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RET"]

[tool.mypy]
strict = true
python_version = "3.11"
plugins = []
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
filterwarnings = ["error"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["spy_trader"]
dev-mode-dirs = ["."]
